[ENH] Refactor away the old NiiLinkExporter
