+++ datman/exporters/bids_legacy.py
@@ -21,15 +21,12 @@ from dcm2bids import dcm2bids, Dcm2bids
 from dcm2bids.sidecar import Acquisition
 
 from datman.exceptions import MetadataException
-from datman.scanid import make_filename
-from datman.utils import (splitext, get_extension, write_json, read_json,
-                          filter_niftis, read_blacklist, get_relative_source,
-                          locate_metadata)
+from datman.utils import (splitext, write_json, read_json, locate_metadata)
 from .base import SessionExporter
 
 logger = logging.getLogger(__name__)
 
-__all__ = ["BidsExporter", "NiiLinkExporter", "BidsOptions"]
+__all__ = ["BidsExporter", "BidsOptions"]
 
 
 class BidsOptions:
@@ -594,356 +591,6 @@ class BidsExporter(SessionExporter):
         return contents
 
 
-class NiiLinkExporter(SessionExporter):
-    """Populates a study's nii folder with symlinks pointing to the bids dir.
-    """
-
-    type = "nii_link"
-    ext = ".nii.gz"
-
-    def __init__(self, config, session, experiment, **kwargs):
-        self.ident = session._ident
-        self.output_dir = session.nii_path
-        self.bids_path = session.bids_path
-        self.config = config
-        self.tags = config.get_tags(site=session.site)
-
-        super().__init__(config, session, experiment, **kwargs)
-
-        self.dm_names = self.get_dm_names()
-        self.bids_names = self.get_bids_niftis()
-        self.name_map = self.match_dm_to_bids(self.dm_names, self.bids_names)
-
-    def get_dm_names(self):
-        """Get the datman-style scan names for an entire XNAT experiment.
-
-        Returns:
-            :obj:`list`: A list of datman-style names for all scans found
-                for the session on XNAT.
-        """
-        names = []
-        for scan in self.experiment.scans:
-            names.extend(scan.names)
-        return names
-
-    def get_bids_niftis(self):
-        """Get all nifti files from a BIDS session.
-
-        Returns:
-            :obj:`list`: A list of full paths (minus the file extension) to
-                each bids format nifti file in the session.
-        """
-        bids_niftis = []
-        for path, _, files in os.walk(self.bids_path):
-            niftis = filter_niftis(files)
-            for item in niftis:
-                basename = item.replace(get_extension(item), "")
-                nii_path = os.path.join(path, basename)
-                if self.belongs_to_session(nii_path):
-                    bids_niftis.append(nii_path)
-        return bids_niftis
-
-    def belongs_to_session(self, nifti_path):
-        """Check if a nifti belongs to this repeat or another for this session.
-
-        Args:
-            nifti_path (str): A nifti file name from the bids folder (minus
-                extension).
-
-        Returns:
-            bool: True if the nifti file belongs to this particular
-                repeat. False if it belongs to another repeat.
-        """
-        try:
-            side_car = read_json(nifti_path + ".json")
-        except FileNotFoundError:
-            # Assume it belongs if a side car cant be read.
-            return True
-
-        repeat = side_car.get("Repeat")
-        if not repeat:
-            # No repeat is recorded in the json, assume its for this session.
-            return True
-
-        return repeat == self.ident.session
-
-    def match_dm_to_bids(self, dm_names, bids_names):
-        """Match each datman file name to its BIDS equivalent.
-
-        Args:
-            dm_names (:obj:`list`): A list of all valid datman scan names found
-                for this session on XNAT.
-            bids_names (:obj:`list`): A list of all bids files (minus
-                extensions) that exist for this session.
-
-        Returns:
-            :obj:`dict`: A dictionary matching the intended datman file name to
-                the full path (minus extension) of the same series in the bids
-                folder. If no matching bids file was found, it will instead be
-                matched to the string 'missing'.
-        """
-        name_map = {}
-        for tag in self.tags:
-            try:
-                bids_conf = self.tags.get(tag)['Bids']
-            except KeyError:
-                logger.info(f"No bids config found for tag {tag}. Can't match "
-                            "bids outputs to a datman-style name.")
-                continue
-
-            matches = self._find_matching_files(bids_names, bids_conf)
-
-            for item in matches:
-                try:
-                    dm_name = self.make_datman_name(item, tag)
-                except Exception as e:
-                    logger.error(
-                        f"Failed to assign datman style name to {item}. "
-                        f"Reason - {e}")
-                    continue
-                name_map[dm_name] = item
-
-        for scan in dm_names:
-            output_file = os.path.join(self.output_dir, scan + self.ext)
-            if scan not in name_map and not os.path.exists(output_file):
-                # An expected scan is missing from the bids folder and
-                # hasnt already been exported directly with dcm2niix
-                name_map[scan] = "missing"
-
-        return name_map
-
-    def make_datman_name(self, bids_path, scan_tag):
-        """Create a Datman-style file name for a bids file.
-
-        Args:
-            bids_path (str): The full path (+/- extension) of a bids file to
-                create a datman name for.
-            scan_tag (str): A datman style tag to apply to the bids scan.
-
-        Returns:
-            str: A valid datman style file name (minus extension).
-        """
-        side_car = read_json(bids_path + ".json")
-        description = side_car['SeriesDescription']
-        num = self.get_series_num(side_car)
-
-        dm_name = make_filename(self.ident, scan_tag, num, description)
-        return dm_name
-
-    def get_series_num(self, side_car):
-        """Find the correct series number for a scan.
-
-        Most JSON side car files have the correct series number already.
-        However, series that are split during nifti conversion (e.g.
-        FMAP-AP/-PA) end up with one of the two JSON files having a modified
-        series number. This function will default to the XNAT series number
-        whenever possible, for accuracy.
-
-        Args:
-            side_car (:obj:`dict`): A dictionary containing the contents of a
-                scan's JSON side car file.
-
-        Returns:
-            str: The most accurate series number found for the scan.
-        """
-        description = side_car['SeriesDescription']
-        num = str(side_car['SeriesNumber'])
-        xnat_scans = [item for item in self.experiment.scans
-                      if item.description == description]
-
-        if not xnat_scans:
-            return num
-
-        if len(xnat_scans) == 1:
-            return xnat_scans[0].series
-
-        # Catch split series (dcm2bids adds 1000 to the series number of
-        # one of the two files)
-        split_num = str(int(num) - 1000).zfill(2)
-        if any([split_num == str(item.series).zfill(2)
-                for item in xnat_scans]):
-            return split_num
-
-        return num
-
-    def _find_matching_files(self, bids_names, bids_conf):
-        """Search a list of bids files to find series that match a datman tag.
-
-        Args:
-            bids_names (:obj:`list`): A list of bids file names to search
-                through.
-            bids_conf (:obj:`dict`): The bids configuration for a single tag
-                from datman's configuration files.
-
-        Returns:
-            :obj:`list`: A list of full paths (minus extension) of bids files
-                that match the tag configuration. If none match, an empty
-                list will be returned.
-        """
-        matches = self._filter_bids(
-            bids_names, bids_conf.get('class'), par_dir=True)
-        matches = self._filter_bids(
-            matches, bids_conf.get(self._get_label_key(bids_conf)))
-        matches = self._filter_bids(matches, bids_conf.get('task'))
-        matches = self._filter_bids(matches, bids_conf.get('dir'))
-        # The below is used to more accurately match FMAP tags
-        matches = self._filter_bids(matches, bids_conf.get('match_acq'))
-        return matches
-
-    def _filter_bids(self, niftis, search_term, par_dir=False):
-        """Find the subset of file names that matches a search string.
-
-        Args:
-            niftis (:obj:`list`): A list of nifti file names to search through.
-            search_term (:obj:`str`): The search term nifti files must match.
-            par_dir (bool, optional): Restricts the search to the nifti file's
-                parent directory, if full paths were given.
-
-        Returns:
-            list: A list of all files that match the search term.
-        """
-        if not search_term:
-            return niftis.copy()
-
-        if not isinstance(search_term, list):
-            search_term = [search_term]
-
-        result = set()
-        for item in niftis:
-            if par_dir:
-                fname = os.path.split(os.path.dirname(item))[1]
-            else:
-                fname = os.path.basename(item)
-
-            for term in search_term:
-                if term in fname:
-                    result.add(item)
-        return list(result)
-
-    def _get_label_key(self, bids_conf):
-        """Return the name for the configuration's label field.
-        """
-        for key in bids_conf:
-            if 'label' in key:
-                return key
-        return ""
-
-    @classmethod
-    def get_output_dir(cls, session):
-        return session.nii_path
-
-    def get_error_file(self, dm_file):
-        return os.path.join(self.output_dir, dm_file + ".err")
-
-    def outputs_exist(self):
-        for dm_name in self.name_map:
-            if read_blacklist(scan=dm_name, config=self.config):
-                continue
-
-            if self.name_map[dm_name] == "missing":
-                if not os.path.exists(self.get_error_file(dm_name)):
-                    return False
-                continue
-
-            full_path = os.path.join(self.output_dir, dm_name + self.ext)
-            if not os.path.exists(full_path):
-                return False
-        return True
-
-    def needs_raw_data(self):
-        return False
-
-    def export(self, *args, **kwargs):
-        # Re run this before exporting, in case new BIDS files exist.
-        self.bids_names = self.get_bids_niftis()
-        self.name_map = self.match_dm_to_bids(self.dm_names, self.bids_names)
-
-        if self.dry_run:
-            logger.info("Dry run: Skipping making nii folder links for "
-                        f"mapping {self.name_map}")
-            return
-
-        if self.outputs_exist():
-            return
-
-        self.make_output_dir()
-        for dm_name, bids_name in self.name_map.items():
-            if bids_name == "missing":
-                self.report_errors(dm_name)
-            else:
-                self.make_link(dm_name, bids_name)
-                # Run in case of previous errors
-                self.clear_errors(dm_name)
-
-    def report_errors(self, dm_file):
-        """Create an error file to report probable BIDS conversion issues.
-
-        Args:
-            dm_file (:obj:`str`): A valid datman file name.
-        """
-        err_file = self.get_error_file(dm_file)
-        contents = (
-            f"{dm_file} could not be made. This may be due to a dcm2bids "
-            "conversion error or an issue with downloading the raw dicoms. "
-            "Please contact an admin as soon as possible.\n"
-        )
-        try:
-            with open(err_file, "w") as fh:
-                fh.write(contents)
-        except Exception as e:
-            logger.error(
-                f"Failed to write error file for {dm_file}. Reason - {e}"
-            )
-
-    def clear_errors(self, dm_file):
-        """Remove an error file from a previous BIDs export issue.
-
-        Args:
-            dm_file (:obj:`str`): A valid datman file name.
-        """
-        err_file = self.get_error_file(dm_file)
-        try:
-            os.remove(err_file)
-        except FileNotFoundError:
-            pass
-        except Exception as e:
-            logger.error(f"Failed while removing {err_file}. Reason - {e}")
-
-    def make_link(self, dm_file, bids_file):
-        """Create a symlink in the datman style that points to a bids file.
-
-        Args:
-            dm_file (:obj:`str`): A valid datman file name.
-            bids_file (:obj:`str`): The full path to a bids file (minus
-                extension.)
-        """
-        base_target = os.path.join(self.output_dir, dm_file)
-        if read_blacklist(scan=base_target, config=self.config):
-            logger.debug(f"Ignoring blacklisted scan {dm_file}")
-            return
-
-        for source in glob(bids_file + '*'):
-            ext = get_extension(source)
-            target = base_target + ext
-
-            if os.path.islink(target) and not os.path.exists(target):
-                # Remove a broken symlink
-                try:
-                    os.unlink(target)
-                except Exception as exc:
-                    logger.error(
-                        f"Failed to remove broken symlink {target} - {exc}")
-                    continue
-
-            rel_source = get_relative_source(source, target)
-            try:
-                os.symlink(rel_source, target)
-            except FileExistsError:
-                pass
-            except Exception as exc:
-                logger.error(f"Failed to create {target}. Reason - {exc}")
-
-
 class FakeSidecar(dcm2bids.Sidecar):
     """Turns XNAT series descriptions into pseudo-sidecars.
     """
@@ -973,16 +620,6 @@ class FakeSidecar(dcm2bids.Sidecar):
 
     def __repr__(self):
         return f"<FakeSidecar {self.data['SeriesDescription']}>"
-
-
-def get_expected_names(participant, sidecars, bids_conf):
-    parser = dcm2bids.SidecarPairing(
-        sidecars, bids_conf["descriptions"]
-    )
-    parser.build_graph()
-    parser.build_acquisitions(participant)
-    parser.find_runs()
-    return [acq.dstRoot for acq in parser.acquisitions]
 
 
 def remove_criteria(descriptions):
